Clarify Client API decomposer and job naming
