+++ nvflare/app_common/executors/client_api/in_process_backend.py
@@ -44,8 +44,8 @@ from nvflare.app_common.app_constant import AppConstants
 from nvflare.app_common.executors.client_api.backend_spec import ClientAPIBackendContext, ClientAPIBackendSpec
 from nvflare.app_common.executors.task_script_runner import TaskScriptRunner
 from nvflare.client.api_spec import CLIENT_API_KEY
-from nvflare.client.cell.decomposers import register_framework_decomposers
 from nvflare.client.config import ConfigKey
+from nvflare.client.decomposers import register_framework_decomposers
 from nvflare.client.in_process.api import (
     TOPIC_ABORT,
     TOPIC_GLOBAL_RESULT,
