+++ tests/integration_test/data/jobs/pt_client_api_in_process_native_tensors/README.md
@@ -1,4 +1,4 @@
-# In-Process PyTorch Tensor Download
+# In-Process Client API Native PyTorch Tensors
 
 This one-round job exercises the inbound half of the in-process framework
 decomposer regression. The server sends a PyTorch model through the streamed
@@ -8,8 +8,8 @@ FOBS path as `TENSOR_DOWNLOAD` (Datum Object Type 6), and the generic
 Run it with:
 
 ```bash
-nvflare simulator tests/integration_test/data/jobs/pt_in_process_dot6 \
-  -w /tmp/nvflare/pt_in_process_dot6 \
+nvflare simulator tests/integration_test/data/jobs/pt_client_api_in_process_native_tensors \
+  -w /tmp/nvflare/pt_client_api_in_process_native_tensors \
   -n 2 \
   -t 2
 ```
