+++ tests/integration_test/data/jobs/pt_client_api_in_process_native_tensors/app/custom/net.py
