+++ tests/integration_test/data/jobs/pt_client_api_in_process_native_tensors/meta.conf
@@ -1,5 +1,5 @@
 {
-  name = "pt_in_process_dot6"
+  name = "pt_client_api_in_process_native_tensors"
   resource_spec {}
   deploy_map {
     app = [
